435: move PackageManager to Platform namespace
