+++ src/SelfManage/BuildTools/BinaryBuildToolFinder.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Php\Pie\SelfManage\BuildTools;
 
+use Php\Pie\Platform\PackageManager;
 use Php\Pie\Platform\TargetPlatform;
 use Symfony\Component\Process\ExecutableFinder;
